# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API para o site Mulheres Palestrantes",
    "version": "1.0.0",
    "title": "Mulheres Palestrantes API"
  },
  "basePath": "/api",
  "tags": [
    {
      "name": "speakers",
      "description": "Gerenciamento de palestrantes"
    },
    {
      "name": "accounts",
      "description": "Gerenciamento de conta de usuário"
    },
    {
      "name": "auth",
      "description": "Autenticação com JWT"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "produces": [
    "application/json",
    "text/plain"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "in": "header",
      "name": "Authentication"
    }
  },
  "paths": {
    "/accounts": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Cria conta de usuário",
        "operationId": "createAccount",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/create_user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "sucesso",
            "schema": {
              "properties": {
                "user": {
                  "$ref": "#/definitions/user"
                }
              }
            }
          },
          "400": {
            "description": "dados inválidos"
          },
          "422": {
            "description": "email não existe ou notificação não pôde ser enviada"
          }
        }
      }
    },
    "/accounts/{id}/confirmation": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Confirma conta de usuário",
        "operationId": "confirmAccount",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "required": true,
            "description": "id da conta do usuário"
          },
          {
            "in": "body",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/confirm_account"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "sucesso",
            "schema": {
              "properties": {
                "user": {
                  "$ref": "#/definitions/confirmation"
                }
              }
            }
          },
          "400": {
            "description": "dados inválidos"
          },
          "422": {
            "description": "conta não pode ser confirmada"
          }
        }
      },
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Envia link de confirmação de usuário por email",
        "operationId": "sendConfirmationLink",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "required": true,
            "description": "id da conta do usuário"
          }
        ],
        "responses": {
          "204": {
            "description": "sucesso"
          },
          "422": {
            "description": "email não pôde ser enviado"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Retorna um token jwt para autenticação no sistema",
        "description": "Apenas usuários que confirmaram sua conta pelo link enviado por email podem se autenticar no sistema",
        "operationId": "authLogin",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "sucesso",
            "schema": {
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/speakers": {
      "get": {
        "tags": [
          "speakers"
        ],
        "summary": "Exibe listagem de palestrantes",
        "operationId": "listSpeakers",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "type": "string",
            "description": "termo de busca por location, name ou interest"
          }
        ],
        "responses": {
          "200": {
            "description": "sucesso",
            "schema": {
              "properties": {
                "speakers": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/speaker"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "speakers"
        ],
        "summary": "Cadastra palestrantes",
        "operationId": "createSpeaker",
        "security": [
          {
            "JWT": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/create_speaker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "sucesso",
            "schema": {
              "properties": {
                "speaker": {
                  "$ref": "#/definitions/speaker"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/speakers/{id}": {
      "get": {
        "tags": [
          "speakers"
        ],
        "summary": "Busca palestrante por id",
        "operationId": "showSpeaker",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "required": true,
            "description": "id do palestrante"
          }
        ],
        "responses": {
          "200": {
            "description": "sucesso",
            "schema": {
              "properties": {
                "speaker": {
                  "$ref": "#/definitions/speaker"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado"
          }
        }
      }
    }
  },
  "definitions": {
    "login": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string",
          "format": "password"
        }
      }
    },
    "speaker": {
      "type": "object",
      "required": [
        "id",
        "name",
        "email",
        "location",
        "photo",
        "site",
        "created_at",
        "updated_at",
        "interests",
        "user_id"
      ],
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "location": {
          "type": "string"
        },
        "photo": {
          "type": "string"
        },
        "site": {
          "type": "string"
        },
        "user_id": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "datetime"
        },
        "updated_at": {
          "type": "string",
          "format": "datetime"
        },
        "interests": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "linkedin": {
          "type": "string"
        },
        "twitter": {
          "type": "string"
        },
        "fb": {
          "type": "string"
        },
        "behance": {
          "type": "string"
        }
      }
    },
    "create_speaker": {
      "type": "object",
      "required": [
        "id",
        "name",
        "email",
        "location",
        "photo",
        "site",
        "created_at",
        "updated_at",
        "interests"
      ],
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "location": {
          "type": "string"
        },
        "photo": {
          "type": "string"
        },
        "site": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "datetime"
        },
        "updated_at": {
          "type": "string",
          "format": "datetime"
        },
        "interests": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "linkedin": {
          "type": "string"
        },
        "twitter": {
          "type": "string"
        },
        "fb": {
          "type": "string"
        },
        "behance": {
          "type": "string"
        }
      }
    },
    "create_user": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "user": {
      "type": "object",
      "required": [
        "email",
        "created_at",
        "updated_at"
      ],
      "properties": {
        "id": {
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "datetime"
        },
        "updated_at": {
          "type": "string",
          "format": "datetime"
        }
      }
    },
    "confirm_account": {
      "type": "object",
      "required": [
        "key"
      ],
      "properties": {
        "key": {
          "type": "string"
        }
      }
    },
    "confirmation": {
      "type": "object",
      "properties": {
        "account_id": {
          "type": "integer"
        },
        "confirmation_date": {
          "type": "string",
          "format": "datetime"
        }
      }
    }
  }
}